+++ java/ip/industrialProcessing/utils/DropBlock.java
@@ -4,7 +4,7 @@ import net.minecraft.entity.item.EntityItem;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.world.World;
-import net.minecraftforge.common.ForgeDirection;
+import net.minecraftforge.common.util.ForgeDirection;
 
 public class DropBlock {
 	public static void doDispense(World par0World, ItemStack par1ItemStack, int par2, ForgeDirection forge, int x, int y, int z) {
@@ -15,12 +15,12 @@ public class DropBlock {
 			double d2 = z;
 			EntityItem entityitem = new EntityItem(par0World, d0, d1 - 0.3D, d2, par1ItemStack);
 			double d3 = par0World.rand.nextDouble() * 0.1D + 0.2D;
-			entityitem.motionX = (double) par3EnumFacing.getFrontOffsetX() * d3;
+			entityitem.motionX = par3EnumFacing.getFrontOffsetX() * d3;
 			entityitem.motionY = 0.20000000298023224D;
-			entityitem.motionZ = (double) par3EnumFacing.getFrontOffsetZ() * d3;
-			entityitem.motionX += par0World.rand.nextGaussian() * 0.007499999832361937D * (double) par2;
-			entityitem.motionY += par0World.rand.nextGaussian() * 0.007499999832361937D * (double) par2;
-			entityitem.motionZ += par0World.rand.nextGaussian() * 0.007499999832361937D * (double) par2;
+			entityitem.motionZ = par3EnumFacing.getFrontOffsetZ() * d3;
+			entityitem.motionX += par0World.rand.nextGaussian() * 0.007499999832361937D * par2;
+			entityitem.motionY += par0World.rand.nextGaussian() * 0.007499999832361937D * par2;
+			entityitem.motionZ += par0World.rand.nextGaussian() * 0.007499999832361937D * par2;
 			par0World.spawnEntityInWorld(entityitem);
 		}
 	}
